Fixes bug with adding new items wrong properties
Was previously not correctly adding the fields. I needed to refactor
anyway so that the displayAddNewForm could be triggered by the edit
button (this is not yet done).

So I refactored to allow displayAddNewForm to create the new todo item, using
the addTodo function, and this solved the bug on adding new items.